+++ work1/work1/include/work1/kernel_vadd.cuh
@@ -2,11 +2,12 @@
 #define _KERNEL_VADD_CUH_
 
 #include <concepts>
+#include <cstdint>
 
 template <std::floating_point ScalarT>
 __global__ void kernel_vadd(
-    const ScalarT* a, const ScalarT* b, ScalarT* c, std::size_t len) {
-  std::size_t i = blockIdx.x * blockDim.x + threadIdx.x;
+    const ScalarT* a, const ScalarT* b, ScalarT* c, std::uint32_t len) {
+  std::uint32_t i = blockIdx.x * blockDim.x + threadIdx.x;
   if (i < len) c[i] = a[i] + b[i];
 }
