+++ work1/work1/include/work1/vadd.hpp
@@ -1,11 +1,11 @@
 #ifndef _WRAPPER_VADD_HPP_
 #define _WRAPPER_VADD_HPP_
 
-#include <cstddef>
+#include <cstdint>
 
 namespace w1 {
 
-void vadd_f32(const float* a, const float* b, float* c, std::size_t len);
+void vadd_f32(const float* a, const float* b, float* c, std::uint32_t len);
 
 }  // namespace w1
